Deleting AssetsManager
